% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_binomial_mixture_bound}
\alias{beta_binomial_mixture_bound}
\title{Beta-binomial uniform boundary.}
\usage{
beta_binomial_mixture_bound(v, alpha, v_opt, g, h, alpha_opt = 0.05,
  is_one_sided = TRUE)
}
\arguments{
\item{v}{intrinsic time value}

\item{alpha}{the error / crossing probability of the boundary}

\item{v_opt}{intrinsic time value for which the boundary is optimized}

\item{g}{lower sub-Bernoulli range parameter}

\item{h}{upper sub-Bernoulli range parameter}

\item{alpha_opt}{alpha for which the boundary is optimized}

\item{is_one_sided}{if FALSE, use the two-sided normal mixture}
}
\description{
Beta-binomial uniform boundary.
}
\examples{
beta_binomial_mixture_bound(c(100, 200), .05, 100, .2, .8)
}
