% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confseq.R
\docType{package}
\name{confseq-package}
\alias{confseq}
\alias{confseq-package}
\title{confseq: Uniform boundaries and confidence sequences}
\description{
This library supports calculation of uniform boundaries, confidence
  sequences, and always-valid p-values. These constructs are useful in sequential
  A/B testing, best-arm identification, and other sequential statistical
  procedures.
}
\details{
The main reference is

Howard, S. R., Ramdas, A., McAuliffe, J., and Sekhon, J. (2018), Uniform,
nonparametric, non-asymptotic confidence sequences, preprint,
https://arxiv.org/abs/1810.08240.

Additionally, the library includes some functions for quantile confidence
sequences and A/B testing based on

Howard, S. R. and Ramdas, A. (2019), Sequential estimation of quantiles with
applications to A/B-testing and best-arm identification, preprint,
https://arxiv.org/abs/1906.09712.

This library is in early-stage development and should not be considered
stable.
}
\author{
\strong{Maintainer}: Steve Howard \email{dev@gostevehoward.com}

}
\keyword{internal}
