% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_poisson_log_mixture}
\alias{gamma_poisson_log_mixture}
\title{Logarithm of mixture supermartingale for the gamma-Poisson mixture.}
\usage{
gamma_poisson_log_mixture(s, v, v_opt, c, alpha_opt = 0.05)
}
\arguments{
\item{s}{value of the underlying martingale}

\item{v}{intrinsic time value}

\item{v_opt}{intrinsic time value for which the corresponding boundary is
optimized}

\item{c}{sub-Poisson scale parameter}

\item{alpha_opt}{alpha for which the corresponding boundary is optimized}
}
\description{
Logarithm of mixture supermartingale for the gamma-Poisson mixture.
}
\examples{
gamma_poisson_log_mixture(c(10, 20), c(100, 400), 100, 2)
}
