% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_stitching_bound}
\alias{poly_stitching_bound}
\title{Polynomial stitched uniform boundary.}
\usage{
poly_stitching_bound(v, alpha, v_min, c = 0, s = 1.4, eta = 2)
}
\arguments{
\item{v}{intrinsic time value}

\item{alpha}{the error / crossing probability of the boundary}

\item{v_min}{intrinsic time for which boundary is optimized}

\item{c}{sub-gamma scale parameter}

\item{s}{controls how crossing probability is distribted over epochs}

\item{eta}{controls the spacing of epochs}
}
\description{
Polynomial stitched uniform boundary.
}
\examples{
poly_stitching_bound(c(100, 200), .05, 100)
}
